pyproject.toml
[tool.poetry]
name = "discord_anime_bot"
version = "0.1.0"
description = ""
authors = ["WildElaeagnus <54314123+akorzunin@users.noreply.github.com>"]

[tool.poetry.dependencies]
python = "^3.10"
pytz = "^2022.1"
youtube_dl = "^2021.12.17"
python-dotenv = "^0.20.0"
requests = "^2.28.1"
nest-asyncio = "^1.5.5"
tinydb = "^4.7.0"
"discord.py" = {git = "https://github.com/Rapptz/discord.py"}
PyNaCl = "^1.5.0"

[tool.poetry.dev-dependencies]
autopep8 = "^1.6.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
